refactor(runtime): extract ai sdk message projection
Generated-by: Codex
